requirements.txt

torch
torchvision
wandb
pandas
Pillow
scikit-learn
matplotlib